Add alerts_helper.rb to remove code duplication
