+++ app/javascript/stimulus_js/controllers/newsletter_email_form_controller.js
@@ -40,22 +40,4 @@ export default class extends Controller {
     button.removeAttribute('disabled');
     button.textContent = 'Preview';
   }
-
-  showAlert(alertType, alertMessage) {
-    const alertWrapper = document.getElementById('flash-messages-wrapper');
-
-    alertWrapper.innerHTML = this.alertTemplate(alertType, alertMessage);
-    window.scrollTo(0, 0);
-  }
-
-  alertTemplate(alertType, alertMessage) {
-    return `<div class="alert alert-${alertType} alert-dismissible fade show flash-message" role="alert">
-              <div class="container-fluid">
-                <center>${alertMessage}</center>
-                <button type="button" class="close" data-dismiss="alert" aria-label="Close">
-                  <span aria-hidden="true">&times;</span>
-                </button>
-              </div>
-            </div>`;
-  }
 }
